Refactoring post_folio code to plugin
