+++ plugins/folio/holdings.py
@@ -1,10 +1,37 @@
-import csv
+import logging
 
 from airflow.models import Variable
 
 from migration_tools.migration_tasks.holdings_csv_transformer import (
     HoldingsCsvTransformer,
 )
+
+from plugins.folio.helpers import post_to_okapi
+
+logger = logging.getLogger(__name__)
+
+def post_folio_holding_records(**kwargs):
+    """Creates/overlays Holdings records in FOLIO"""
+    dag = kwargs["dag_run"]
+
+    batch_size = kwargs.get("MAX_ENTITIES", 1000)
+    job_number = kwargs.get("job")
+
+    with open(f"/tmp/holdings-{dag.run_id}-{job_number}.json") as fo:
+        holding_records = json.load(fo)
+
+    for i in range(0, len(holding_records), batch_size):
+        holdings_batch = holding_records[i: i + batch_size]
+        logger.info(f"Posting {i} to {i+batch_size} holding records")
+        _post_to_okapi(
+            token=kwargs["task_instance"].xcom_pull(
+                key="return_value", task_ids="post-to-folio.folio_login"
+            ),
+            records=holdings_batch,
+            endpoint="/holdings-storage/batch/synchronous?upsert=true",
+            payload_key="holdingsRecords",
+            **kwargs,
+        )
 
 
 def run_holdings_tranformer(*args, **kwargs):
