+++ plugins/folio/instances.py
@@ -1,0 +1,60 @@
+import logging
+
+from migration_tools.migration_tasks.bibs_transformer import BibsTransformer
+
+from plugins.folio.helpers import post_to_okapi
+
+logger = logging.getLogger(__name__)
+
+def post_folio_instance_records(**kwargs):
+    """Creates new records in FOLIO"""
+    dag = kwargs["dag_run"]
+
+    batch_size = kwargs.get("MAX_ENTITIES", 1000)
+    job_number = kwargs.get("job")
+
+    with open(f"/tmp/instances-{dag.run_id}-{job_number}.json") as fo:
+        instance_records = json.load(fo)
+
+    for i in range(0, len(instance_records), batch_size):
+        instance_batch = instance_records[i: i + batch_size]
+        logger.info(f"Posting {len(instance_batch)} in batch {i/batch_size}")
+        post_to_okapi(
+            token=kwargs["task_instance"].xcom_pull(
+                key="return_value", task_ids="post-to-folio.folio_login"
+            ),
+            records=instance_batch,
+            endpoint="/instance-storage/batch/synchronous?upsert=true",
+            payload_key="instances",
+            **kwargs,
+        )
+
+
+def run_bibs_transformer(*args, **kwargs):
+    dag = kwargs["dag_run"]
+
+    library_config = kwargs["library_config"]
+
+    marc_stem = kwargs["marc_stem"]
+    
+    library_config.iteration_identifier = dag.run_id
+
+    bibs_configuration = BibsTransformer.TaskConfiguration(
+        name="bibs-transformer",
+        migration_task_type="BibsTransformer",
+        library_config=library_config,
+        hrid_handling="default",
+        files=[{ "file_name": f"{marc_stem}.mrc", "suppress": False}],
+        ils_flavour="voyager",  # Voyager uses 001 field, using tag001 works
+    )
+
+    bibs_transformer = BibsTransformer(
+        bibs_configuration, library_config, use_logging=False
+    )
+
+    logger.info(f"Starting bibs_tranfers work for {marc_stem}.mrc")
+
+    bibs_transformer.do_work()
+
+    bibs_transformer.wrap_up()
+
