+++ plugins/folio/login.py
@@ -1,0 +1,21 @@
+import requests
+
+from airflow.models import Variable
+
+def folio_login(**kwargs):
+    """Logs into FOLIO and returns Okapi token."""
+    okapi_url = Variable.get("OKAPI_URL")
+    username = Variable.get("FOLIO_USER")
+    password = Variable.get("FOLIO_PASSWORD")
+    tenant = "sul"
+
+    data = {"username": username, "password": password}
+    headers = {"Content-type": "application/json", "x-okapi-tenant": tenant}
+
+    url = f"{okapi_url}/authn/login"
+    result = requests.post(url, json=data, headers=headers)
+
+    if result.status_code == 201:  # Valid token created and returned
+        return result.headers.get("x-okapi-token")
+
+    result.raise_for_status()
