+++ plugins/tests/test_helpers.py
@@ -1,0 +1,4 @@
+from plugins.folio.helpers import tranform_csv_to_tsv
+
+def test_tranform_csv_to_tsv():
+    assert tranform_csv_to_tsv
